parse_view_column_config: avoid reading uninitialized/NULL variable
If the view does not support a specific type or we cannot
find a matching column, we do not want to use "column->type"
to report the error. It is either uninitialized or NULL in
these error conditions. Instead, we can want to use the type
we are looking for.